SC-3111: Benchmark 1.0 refactoring
