--- bootstrap/Yves/bootstrap.php
@@ -1,19 +1,0 @@
-<?php
-
-use Spryker\Shared\Config\Application\Environment;
-use Spryker\Shared\ErrorHandler\ErrorHandlerEnvironment;
-use SprykerSdk\Yves\Benchmark\Bootstrap\FallbackYvesBootstrap;
-
-define('APPLICATION', 'YVES');
-defined('APPLICATION_ROOT_DIR') || define('APPLICATION_ROOT_DIR', realpath(__DIR__ . '/../../../../../..'));
-
-require_once APPLICATION_ROOT_DIR . '/vendor/autoload.php';
-
-Environment::initialize();
-
-$errorHandlerEnvironment = new ErrorHandlerEnvironment();
-$errorHandlerEnvironment->initialize();
-
-$bootstrap = new FallbackYvesBootstrap();
-
-$bootstrap->boot();
