+++ bootstrap/bootstrap.php
@@ -1,0 +1,10 @@
+<?php
+
+use Spryker\Shared\Config\Application\Environment;
+
+define('APPLICATION', 'ZED');
+defined('APPLICATION_ROOT_DIR') || define('APPLICATION_ROOT_DIR', realpath(__DIR__ . '/../../../..'));
+
+require_once APPLICATION_ROOT_DIR . '/vendor/autoload.php';
+
+Environment::initialize();
