+++ src/SprykerSdk/Glue/Benchmark/BenchmarkConfig.php
@@ -10,9 +10,6 @@ namespace SprykerSdk\Glue\Benchmark;
 use Spryker\Yves\Kernel\AbstractBundleConfig;
 use SprykerSdk\Shared\Benchmark\BenchmarkConstants;
 
-/**
- * @method \SprykerSdk\Shared\Benchmark\BenchmarkConfig getSharedConfig()
- */
 class BenchmarkConfig extends AbstractBundleConfig
 {
     /**
@@ -23,45 +20,5 @@ class BenchmarkConfig extends AbstractBundleConfig
     public function getRequestBaseUrl(): string
     {
         return $this->get(BenchmarkConstants::GLUE_APPLICATION_DOMAIN);
-    }
-
-    /**
-     * @api
-     *
-     * @return string
-     */
-    public function getTestsDirectory(): string
-    {
-        return $this->getSharedConfig()->getTestsDirectory();
-    }
-
-    /**
-     * @api
-     *
-     * @return int
-     */
-    public function getIterations(): int
-    {
-        return $this->getSharedConfig()->getDefaultIterationsConfig();
-    }
-
-    /**
-     * @api
-     *
-     * @return int
-     */
-    public function getRevolutions(): int
-    {
-        return $this->getSharedConfig()->getDefaultRevolutionsConfig();
-    }
-
-    /**
-     * @api
-     *
-     * @return string
-     */
-    public function getReport(): string
-    {
-        return $this->getSharedConfig()->getDefaultReportConfig();
     }
 }
