--- src/SprykerSdk/Glue/Benchmark/Bootstrap/FallbackGlueBootstrap.php
@@ -1,14 +1,0 @@
-<?php
-
-/**
- * Copyright © 2016-present Spryker Systems GmbH. All rights reserved.
- * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
- */
-
-namespace SprykerSdk\Glue\Benchmark\Bootstrap;
-
-use Spryker\Glue\GlueApplication\Bootstrap\GlueBootstrap;
-
-class FallbackGlueBootstrap extends GlueBootstrap
-{
-}
