--- src/SprykerSdk/Glue/Benchmark/PhpBench/PhpBenchRunner.php
@@ -1,14 +1,0 @@
-<?php
-
-/**
- * Copyright © 2016-present Spryker Systems GmbH. All rights reserved.
- * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
- */
-
-namespace SprykerSdk\Glue\Benchmark\PhpBench;
-
-use SprykerSdk\Shared\Benchmark\PhpBench\AbstractPhpBenchRunner;
-
-class PhpBenchRunner extends AbstractPhpBenchRunner
-{
-}
