+++ src/SprykerSdk/Zed/Benchmark/Business/BenchmarkBusinessFactory.php
@@ -10,18 +10,18 @@ namespace SprykerSdk\Zed\Benchmark\Business;
 use GuzzleHttp\Cookie\CookieJarInterface;
 use Spryker\Zed\Kernel\Business\AbstractBusinessFactory;
 use SprykerSdk\Client\Benchmark\BenchmarkClientInterface;
-use SprykerSdk\Shared\Benchmark\Command\CommandBuilderInterface;
 use SprykerSdk\Shared\Benchmark\Helper\CsrfToken\CsrfTokenHelperInterface;
 use SprykerSdk\Shared\Benchmark\Helper\CsrfToken\FormCsrfTokenHelper;
 use SprykerSdk\Shared\Benchmark\Helper\Http\HttpHelper;
 use SprykerSdk\Shared\Benchmark\Helper\Http\HttpHelperInterface;
 use SprykerSdk\Shared\Benchmark\Helper\Login\LoginHelperInterface;
-use SprykerSdk\Shared\Benchmark\PhpBench\PhpBenchRunnerInterface;
 use SprykerSdk\Shared\Benchmark\Request\RequestBuilderInterface;
 use SprykerSdk\Zed\Benchmark\BenchmarkDependencyProvider;
 use SprykerSdk\Zed\Benchmark\Business\Command\CommandBuilder;
+use SprykerSdk\Zed\Benchmark\Business\Command\CommandBuilderInterface;
 use SprykerSdk\Zed\Benchmark\Business\Helper\Login\LoginHelper;
 use SprykerSdk\Zed\Benchmark\Business\PhpBench\PhpBenchRunner;
+use SprykerSdk\Zed\Benchmark\Business\PhpBench\PhpBenchRunnerInterface;
 use SprykerSdk\Zed\Benchmark\Business\Request\RequestBuilder;
 use SprykerSdk\Zed\Benchmark\Dependency\Service\BenchmarkToUtilEncodingServiceInterface;
 
@@ -31,11 +31,19 @@ use SprykerSdk\Zed\Benchmark\Dependency\Service\BenchmarkToUtilEncodingServiceIn
 class BenchmarkBusinessFactory extends AbstractBusinessFactory
 {
     /**
-     * @return \SprykerSdk\Shared\Benchmark\PhpBench\PhpBenchRunnerInterface
+     * @return \SprykerSdk\Zed\Benchmark\Business\PhpBench\PhpBenchRunnerInterface
      */
     public function createPhpBenchRunner(): PhpBenchRunnerInterface
     {
         return new PhpBenchRunner($this->createCommandBuilder());
+    }
+
+    /**
+     * @return \SprykerSdk\Zed\Benchmark\Business\Command\CommandBuilderInterface
+     */
+    public function createCommandBuilder(): CommandBuilderInterface
+    {
+        return new CommandBuilder($this->getConfig());
     }
 
     /**
@@ -97,13 +105,5 @@ class BenchmarkBusinessFactory extends AbstractBusinessFactory
     public function getUtilEncodingService(): BenchmarkToUtilEncodingServiceInterface
     {
         return $this->getProvidedDependency(BenchmarkDependencyProvider::SERVICE_UTIL_ENCODING);
-    }
-
-    /**
-     * @return \SprykerSdk\Shared\Benchmark\Command\CommandBuilderInterface
-     */
-    public function createCommandBuilder(): CommandBuilderInterface
-    {
-        return new CommandBuilder($this->getConfig());
     }
 }
