--- src/SprykerSdk/Zed/Benchmark/Business/Bootstrap/FallbackZedBootstrap.php
@@ -1,14 +1,0 @@
-<?php
-
-/**
- * Copyright © 2016-present Spryker Systems GmbH. All rights reserved.
- * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
- */
-
-namespace SprykerSdk\Zed\Benchmark\Business\Bootstrap;
-
-use Spryker\Zed\Application\Communication\ZedBootstrap;
-
-class FallbackZedBootstrap extends ZedBootstrap
-{
-}
