--- src/SprykerSdk/Zed/Benchmark/Communication/BenchmarkCommunicationFactory.php
@@ -1,18 +1,0 @@
-<?php
-
-/**
- * Copyright © 2016-present Spryker Systems GmbH. All rights reserved.
- * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
- */
-
-namespace SprykerSdk\Zed\Benchmark\Communication;
-
-use Spryker\Zed\Kernel\Communication\AbstractCommunicationFactory;
-
-/**
- * @method \SprykerSdk\Zed\Benchmark\BenchmarkConfig getConfig()
- * @method \SprykerSdk\Zed\Benchmark\Business\BenchmarkFacadeInterface getFacade()
- */
-class BenchmarkCommunicationFactory extends AbstractCommunicationFactory
-{
-}
